status infos refactored
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@3797 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/server/restful/servlets/rtmp/JoystickServlet.java
@@ -45,7 +45,7 @@ public class JoystickServlet extends JSIDPlay2Servlet {
 		try {
 			UUID uuid = UUID.fromString(request.getParameter("name"));
 			int number = Integer.valueOf(request.getParameter("number"));
-			byte value = Byte.valueOf(request.getParameter("value"));
+			int value = Integer.valueOf(request.getParameter("value"));
 
 			info(String.format("joystick: RTMP stream of: %s, number=%d, value=%d", uuid, number, value));
 			update(uuid, rtmpPlayerWithStatus -> rtmpPlayerWithStatus.joystick(number, value));
